repo update

+++ code/_onclick/telekinesis.dm
@@ -71,7 +71,7 @@
 	icon = 'icons/obj/magic.dmi'//Needs sprites
 	icon_state = "2"
 	item_flags = NOBLUDGEON | ABSTRACT | DROPDEL
-	//item_state = null
+	//inhand_icon_state = null
 	w_class = WEIGHT_CLASS_GIGANTIC
 	layer = ABOVE_HUD_LAYER
 	plane = ABOVE_HUD_PLANE
@@ -169,7 +169,7 @@
 	return TRUE
 
 /obj/item/tk_grab/proc/check_if_focusable(obj/target)
-	if(!tk_user || !istype(tk_user) || QDELETED(target) || !istype(target) || !tk_user.dna.check_mutation(TK))
+	if(!tk_user || !istype(tk_user) || QDELETED(target) || !istype(target) || (!tk_user.dna.check_mutation(TK) && !tk_user.mind.has_martialart(MARTIALART_STARTERSITH) ) ) //FULPSTATION Chaplain Starter Sith PR by Surrealistik Jan 2020; allows Sith telekinesis
 		qdel(src)
 		return
 	if(!tkMaxRangeCheck(tk_user, target) || target.anchored || !isturf(target.loc))
